Extract shift/rotate instructions to instructions_shift.go module
